+++ projects/ccd-case-ui-toolkit/src/lib/app-config.mock.ts
@@ -154,6 +154,10 @@ export class AppMockConfig implements AbstractAppConfig {
   }
 
   public getPaymentReturnUrl(): string {
-      return '';
+    return '';
+  }
+
+  public getCategoriesAndDocumentsUrl(): string {
+    return '';
   }
 }