+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/case-file-view/index.ts
@@ -1,0 +1,3 @@
+export * from './case-file-view-category.model';
+export * from './case-file-view-document.model';
+export * from './categories-and-documents.model';